moving HaikuReturn into separate components
