+++ src/components/DisplayHaiku.js
@@ -1,4 +1,5 @@
 import {useState} from 'react';
+import HaikuReturn from './HaikuReturn';
 
 
 const DisplayHaiku = (props) => {
@@ -23,48 +24,38 @@ const DisplayHaiku = (props) => {
     // }
 
     return(
-        <>
-            <h2>Here's Your Haiku!</h2>
-            <ul 
-                className="line1"
-                style= {syllables >=5 ? {color:"grey"} : {color:"black"}}
-            >
-                {haikuLines[0].map((word)=>{
-                    return(
-                    <li key={word.key}>{word.word}</li>
-                    )
-                })
-                }
-                {/* <button
-                    // style={syllables <= 5 ? {display:"block"} : {display:"none"}}
-                    onClick={handleClick} >REMOVE</button> */}
+        // <div className='haikuBox'>
+        //     <ul 
+        //         className="line1 haikuLine"
+        //         style= {syllables >=5 ? {color:"grey"} : {color:"black"}}
+        //     >
+        //         {
+        //             liFromArray(haikuLines, 0)
+        //         }
+        //         {/* <button
+        //             // style={syllables <= 5 ? {display:"block"} : {display:"none"}}
+        //             onClick={handleClick} >REMOVE</button> */}
                 
-            </ul>
-            <ul
-                className="line2"
-                style={syllables >= 12 ? { color: "grey" } : { color: "black" }}
-            >
-                {haikuLines[1].map((word) => {
-                    return(
-                    <li key={word.key}>{word.word}</li>
-                    )
-                })
-                }
-            </ul>
-            <ul
-                className="line3"
-                style={syllables == 17 ? { color: "grey" } : { color: "black" }}
-            >
-                {haikuLines[2].map((word) => {
-                    return(
-                    <li key={word.key}>{word.word}</li>
-                    )
-                })
-                }
-            </ul>
-            {/* <button onClick={handleClick}>Not happy with a word? get rid of it</button> */}
-        </>
-        
+        //     </ul>
+        //     <ul
+        //         className="line2 haikuLine"
+        //         style={syllables >= 12 ? { color: "grey" } : { color: "black" }}
+        //     >
+        //         {
+        //             liFromArray(haikuLines, 1)
+        //         }
+        //     </ul>
+        //     <ul
+        //         className="line3 haikuLine"
+        //         style={syllables == 17 ? { color: "grey" } : { color: "black" }}
+        //     >
+        //         {
+        //             liFromArray(haikuLines, 2)
+        //         }
+        //     </ul>
+        //     {/* <button onClick={handleClick}>Not happy with a word? get rid of it</button> */}
+        // </div>
+        <HaikuReturn syllables={ syllables } haikuArray={ haikuLines } />
     )
 }
