+++ src/reusableLogic/liFromArray.js
@@ -1,0 +1,12 @@
+
+const liFromArray = (array, index) => {
+    return(
+        array[index].map((word) => {
+            return (
+                <li key={word.key}>{word.word}</li>
+            )
+        })
+    )
+}
+
+export default liFromArray
